Adding code contributed by aventurella:
CHANGELOG

Added the ability to send multipart/form-data (includes the ability to
upload files) (Content-Type header must be set to multipart/form-data)
Added the ability to specify key/value pairs for
application/x-www-form-urlencoded forms instead of typing the raw
input. (Content-Type must be set to application/x-www-form-urlencoded)
Added the ability to swap between raw input and key/value pair entry.
Added the ability to select raw input, if no raw input is entered, and
the request method is POST or PUT and a file is available in the file
table, a CRCFileRequest will be initiated. (No views to queue the user
this is happening, but it's setup to work that way)

Externalized Request building into their own objects:
CRCMultipartRequest, CRCFormEncodedRequest, CRCRawRequest,
CRCFileRequest
Externalized saving requests into CRCRequest objects instead of an NSDictionary